requirements.txt
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.21.0
pillow>=9.5.0
opencv-python>=4.7.0
accelerate>=0.20.0
safetensors>=0.3.0

